+++ src/main/java/com/boje/hotelbooking/serviceImp/UserServiceImp.java
@@ -29,29 +29,27 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    public UserDTORequest updateUser(UserDTORequest userDTORequest) {
-        User user = userRepository.findById(userDTORequest.getId())
-                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + userDTORequest.getId()));
+    public UserDTORequest updateUser(int user_id, UserDTO userDTO) {
+        User user = userRepository.findById(user_id)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + user_id));
 
-        user.setAge(userDTORequest.getAge());
-        user.setBookings(userDTORequest.getBookings());
-        user.setPassword(userDTORequest.getPassword());
-        user.setFullName(userDTORequest.getFullName());
-        user.setContactInfo(userDTORequest.getContactInfo());
+        user.setAge(userDTO.getAge());
+        user.setBookings(userDTO.getBookings());
+        user.setPassword(userDTO.getPassword());
+        user.setFullName(userDTO.getFullName());
+        user.setContactInfo(userDTO.getContactInfo());
 
         User updatedUser = userRepository.save(user);
         return entityMapper.toUserDTORequest(updatedUser);
     }
 
     @Override
-    public boolean deleteUser(int user_id) {
-        try{
-            userRepository.deleteById(user_id);
-            return true;
+    public void deleteUser(int user_id) {
+        if (!userRepository.existsById(user_id)) {
+            throw new ResourceNotFoundException("Contact info not found with id: " + user_id);
         }
-        catch(Exception e){
-            return false;
-        }
+
+        userRepository.deleteById(user_id);
     }
 
     @Override
